requirements.txt
azure-functions
pymongo==3.12.3
azure-communication-callautomation==1.2.0
grpcio==1.62.0
grpcio-status==1.62.0

